(PC-22272) feat(ui): extract ButtonInsideText tests from _tests_ folder
